+++ src/agent/base.ts
@@ -3,10 +3,10 @@ import os from "os";
 import {AIMessage, AIMessageChunk, BaseMessage, HumanMessage, SystemMessage} from "@langchain/core/messages";
 import {DynamicStructuredTool} from "@langchain/core/tools";
 import {convertJSONSchemaDraft7ToZod} from "../shared/util/draftToZod";
-import {MCPClient} from "./client";
+import {MCPClient} from "./mcp.client";
 import {ToolNode} from "./util/tool-node";
 import {Annotation, MemorySaver, messagesStateReducer, StateGraph} from "@langchain/langgraph";
-import {createViewImageTool} from "./local_tools/image_tool";
+import {createViewImageTool} from "./local_tools/image.tool";
 import {StreamEvent} from "./types";
 import {SystemMessagePromptTemplate} from "@langchain/core/prompts";
 import { MessageContentText } from "@langchain/core/dist/messages/base";
