+++ src/agent/util/tool-node.ts
@@ -6,7 +6,7 @@ import {isLocalTool} from "../local_tools/base";
 import {Logger} from '../../logger';
 import {z} from 'zod';
 import {zodToJsonSchema} from 'zod-to-json-schema';
-import {StateAnnotation} from "../baseAgent";
+import {StateAnnotation} from "../base";
 
 export type ToolNodeOptions = {
   name?: string;
